+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import EventModel, { Event, eventDAO, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -61,6 +61,11 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "Event not found.")
+            }
+
             const result = await EventModel.findById(id);
             return response.success(res, result, "Event found successfully!");
         } catch (error) {
@@ -70,6 +75,11 @@ export default {
     async findOneBySlug(req: IReqUser, res: Response) {
         try {
             const { slug } = req.params;
+
+            if (!isValidObjectId(slug)) {
+                return response.notFound(res, "Event not found.");
+            }
+
             const result = await EventModel.findOne({
                 slug: slug
             });
@@ -81,6 +91,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "Event not found.");
+            }
+
             const result = await EventModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -92,6 +107,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "Event not found.");
+            }
+
             const result = await EventModel.findByIdAndDelete(id, {
                     new: true
                 });
